Make component list reusable
